+++ WaterSupply/app/src/main/java/com/example/watersupply/mainActivityCalls/countDownTimer/timerFunctionality/TimerFunctionality.kt
@@ -10,11 +10,9 @@ import com.example.watersupply.SharedPreferencesInitialization
  */
 class TimerFunctionality {
 
-
     private val sharedPreferencesInitialization = SharedPreferencesInitialization()
 
-
-    var convertedTimerCountToHours: Long = sharedPreferencesInitialization.timerSeekBarValueOutput * 3600
+    var convertedTimerCountToHours: Long = (sharedPreferencesInitialization.timerSeekBarValueOutput * 3600).toLong()
     var detectTimeChange : Long = 0L
 
     var secondsRemaining = convertedTimerCountToHours
@@ -22,6 +20,4 @@ class TimerFunctionality {
 
     var timerState = TimerState.Paused
     lateinit var countDownTimer: CountDownTimer
-
-
 }
